Roll extensions into separate file
